Move filters to separate folder
