+++ pose-detection/src/calculators/interfaces/config_interfaces.ts
