+++ pose-detection/src/calculators/keypoint_tracker.ts
@@ -15,9 +15,9 @@
  * =============================================================================
  */
 
-import {Keypoint} from '../../shared/types';
+import {Keypoint} from '../shared/calculators/interfaces/common_interfaces';
 
-import {Pose} from '../../types';
+import {Pose} from '../types';
 
 import {Track} from './interfaces/common_interfaces';
 import {TrackerConfig} from './interfaces/config_interfaces';
