+++ pose-detection/src/calculators/types.ts
