+++ pose-detection/src/movenet/crop_utils.ts
@@ -16,9 +16,8 @@
  */
 
 import {COCO_KEYPOINTS} from '../constants';
-import {ImageSize} from '../shared/calculators/interfaces/common_interfaces';
+import {ImageSize, Keypoint} from '../shared/calculators/interfaces/common_interfaces';
 import {BoundingBox} from '../shared/calculators/interfaces/shape_interfaces';
-import {Keypoint} from '../shared/types';
 
 import {MIN_CROP_KEYPOINT_SCORE} from './constants';
