+++ pose-detection/src/posenet/calculators/decode_multiple_poses_util.ts
@@ -17,7 +17,7 @@
 import * as tf from '@tensorflow/tfjs-core';
 
 import {COCO_KEYPOINTS} from '../../constants';
-import {Keypoint} from '../../shared/types';
+import {Keypoint} from '../../shared/calculators/interfaces/common_interfaces';
 import {Pose} from '../../types';
 import {NUM_KEYPOINTS, POSE_CHAIN} from '../constants';
 import {NumberDict, NumberTuple, Part, PartWithScore, Vector2D} from '../types';
