+++ pose-detection/src/posenet/calculators/decode_single_pose.ts
@@ -18,7 +18,7 @@
 import * as tf from '@tensorflow/tfjs-core';
 
 import {COCO_KEYPOINTS} from '../../constants';
-import {Keypoint} from '../../shared/types';
+import {Keypoint} from '../../shared/calculators/interfaces/common_interfaces';
 import {Pose} from '../../types';
 import {PoseNetOutputStride} from '../types';
