+++ pose-detection/src/types.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  * =============================================================================
  */
+import {DetectorInput, Keypoint} from './shared/calculators/interfaces/common_interfaces';
 import {BoundingBox} from './shared/calculators/interfaces/shape_interfaces';
-import {DetectorInput, Keypoint} from './shared/types';
+
 export {Keypoint};
 
 export enum SupportedModels {
