+++ shared/calculators/calculate_score_copy.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {Keypoint} from '../types';
+import {Keypoint} from './interfaces/common_interfaces';
 
 /**
  * A calculator to copy score between landmarks.
