+++ shared/calculators/refine_landmarks_from_heatmap.ts
@@ -16,7 +16,7 @@
  */
 import * as tf from '@tensorflow/tfjs-core';
 
-import {Keypoint} from '../types';
+import {Keypoint} from './interfaces/common_interfaces';
 
 import {RefineLandmarksFromHeatmapConfig} from './interfaces/config_interfaces';
