+++ shared/calculators/remove_landmark_letterbox.ts
@@ -15,8 +15,7 @@
  * =============================================================================
  */
 
-import {Keypoint} from '../types';
-import {Padding} from './interfaces/common_interfaces';
+import {Keypoint, Padding} from './interfaces/common_interfaces';
 
 /**
  * Adjusts landmark locations on a letterboxed image to the corresponding
