+++ shared/filters/keypoints_velocity_filter.ts
@@ -14,28 +14,27 @@
  * limitations under the License.
  * =============================================================================
  */
+import {Keypoint, KeypointsFilter} from '../calculators/interfaces/common_interfaces';
+import {VelocityFilterConfig} from '../calculators/interfaces/config_interfaces';
+import {RelativeVelocityFilter} from './relative_velocity_filter';
 
-import {Keypoint} from '../types';
 
-import {KeypointsFilter} from './interfaces/common_interfaces';
-import {OneEuroFilterConfig} from './interfaces/config_interfaces';
-import {OneEuroFilter} from './one_euro_filter';
 
 /**
- * A stateful filter that smoothes keypoints values overtime.
+ * A stateful filter that smoothes landmark values overtime.
  *
- * More specifically, it uses `OneEuroFilter` to smooth every x, y, z
+ * More specifically, it uses `RelativeVelocityFilter` to smooth every x, y, z
  * coordinates over time, which as result gives us velocity of how these values
  * change over time. With higher velocity it weights new values higher.
  */
 // ref:
 // https://github.com/google/mediapipe/blob/master/mediapipe/calculators/util/landmarks_smoothing_calculator.cc
-export class KeypointsOneEuroFilter implements KeypointsFilter {
-  private xFilters: OneEuroFilter[];
-  private yFilters: OneEuroFilter[];
-  private zFilters: OneEuroFilter[];
+export class KeypointsVelocityFilter implements KeypointsFilter {
+  private xFilters: RelativeVelocityFilter[];
+  private yFilters: RelativeVelocityFilter[];
+  private zFilters: RelativeVelocityFilter[];
 
-  constructor(private readonly config: OneEuroFilterConfig) {}
+  constructor(private readonly config: VelocityFilterConfig) {}
 
   apply(keypoints: Keypoint[], microSeconds: number, objectScale: number):
       Keypoint[] {
@@ -43,10 +42,6 @@ export class KeypointsOneEuroFilter implements KeypointsFilter {
       this.reset();
       return null;
     }
-
-    // Initialize filters once.
-    this.initializeFiltersIfEmpty(keypoints);
-
     // Get value scale as inverse value of the object scale.
     // If value is too small smoothing will be disabled and keypoints will be
     // returned as is.
@@ -55,8 +50,11 @@ export class KeypointsOneEuroFilter implements KeypointsFilter {
       if (objectScale < this.config.minAllowedObjectScale) {
         return [...keypoints];
       }
-      valueScale = 1.0 / objectScale;
+      valueScale = 1 / objectScale;
     }
+
+    // Initialize filters once.
+    this.initializeFiltersIfEmpty(keypoints);
 
     // Filter keypoints. Every axis of every keypoint is filtered separately.
     return keypoints.map((keypoint, i) => {
@@ -85,9 +83,12 @@ export class KeypointsOneEuroFilter implements KeypointsFilter {
   // check the size.
   private initializeFiltersIfEmpty(keypoints: Keypoint[]) {
     if (this.xFilters == null || this.xFilters.length !== keypoints.length) {
-      this.xFilters = keypoints.map(_ => new OneEuroFilter(this.config));
-      this.yFilters = keypoints.map(_ => new OneEuroFilter(this.config));
-      this.zFilters = keypoints.map(_ => new OneEuroFilter(this.config));
+      this.xFilters =
+          keypoints.map(_ => new RelativeVelocityFilter(this.config));
+      this.yFilters =
+          keypoints.map(_ => new RelativeVelocityFilter(this.config));
+      this.zFilters =
+          keypoints.map(_ => new RelativeVelocityFilter(this.config));
     }
   }
 }
