+++ shared/filters/low_pass_filter.ts
